backend/requirements.txt
fastapi==0.95.1
motor==3.1.2
pymongo==4.3.3
python-dotenv==1.0.0
pydantic==1.10.7
passlib==1.7.4
uvicorn==0.21.1
bcrypt==4.0.1
motor==3.1.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
typer==0.8.0






